refactor(script): split NPC buildins into buildin_npc.c
Issue #61 third step: move NPC control, timers, waitingroom, and walk/shop helpers out of script.c; share sleep/awake APIs via script_internal.h.
