+++ vcproj/map-server.vcxproj.filters
@@ -109,6 +109,9 @@
     <ClCompile Include="..\src\map\buildin_map.c">
       <Filter>map_sql</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\map\buildin_npc.c">
+      <Filter>map_sql</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\map\skill.c">
       <Filter>map_sql</Filter>
     </ClCompile>
